requirement.txt
asyncssh==1.18.0
bcrypt==3.1.7
cffi==1.13.2
cryptography==2.8
future==0.18.2
Jinja2==2.10.3
MarkupSafe==1.1.1
netdev==0.9.3
netmiko==3.0.0
paramiko==2.7.1
pycparser==2.19
PyNaCl==1.3.0
pyserial==3.4
PyYAML==5.3
scp==0.13.2
six==1.14.0
textfsm==1.1.0
